+++ app/Http/Controllers/CoordenacaoController.php
@@ -16,9 +16,13 @@ class CoordenacaoController extends Controller
 {
     public function index()
     {
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }
 
         //Pegando informações do usuário que está acessando o sistema
-        $matricula  =   getenv('USERNAME');
+        $matricula = session('matricula');
 
         //Listando as empresas
         $Empresas = Empresa::all();
@@ -47,7 +51,10 @@ class CoordenacaoController extends Controller
 
     public function editar($id)
     {
-        
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }
         //Pegando informações do usuário que está acessando o sistema 
         $matricula  =  getenv('USERNAME');
 
@@ -75,6 +82,10 @@ class CoordenacaoController extends Controller
 
     public function incluir(Request $request) 
     {
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }
 
         //Verificando se a solicitação veio da inclusão ou edição
         if($request->input('idcoordenacao') == "") { 
@@ -102,6 +113,10 @@ class CoordenacaoController extends Controller
 
     public function delete($id)
     {
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }   
         //Encontrando e deletando Coordenacao (softDelete)
         $Coordenacao = Coordenacao::find($id);
         $Coordenacao->delete();
