+++ app/Http/Controllers/GestorController.php
@@ -20,9 +20,13 @@ class GestorController extends Controller
 {
     public function index()
     {
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }
 
         //Pegando informações do usuário que está acessando o sistema
-        $matricula  =   getenv('USERNAME');
+        $matricula = session('matricula');
 
         //Listando as empresas
         $Empresas = Empresa::all();
@@ -56,6 +60,10 @@ class GestorController extends Controller
 
     public function editar($id)
     {
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }
         
         //Pegando informações do usuário que está acessando o sistema 
         $matricula  =  getenv('USERNAME');
@@ -88,6 +96,10 @@ class GestorController extends Controller
 
     public function incluir(Request $request) 
     {
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }
 
         //Verificando se a solicitação veio da inclusão ou edição
         if($request->input('idgestor') == "") { 
@@ -112,6 +124,10 @@ class GestorController extends Controller
 
     public function delete($id)
     {
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }
         //Encontrando e deletando Contrato (softDelete)
         $Gestor = Gestor::find($id);
         $Gestor->delete();
