+++ app/Http/Controllers/MacrocelulaController.php
@@ -18,9 +18,13 @@ class MacrocelulaController extends Controller
 {
     public function index()
     {
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }
 
         //Pegando informações do usuário que está acessando o sistema
-        $matricula  =   getenv('USERNAME');
+        $matricula = session('matricula');
 
         //Listando as empresas
         $Empresas = Empresa::all();
@@ -49,6 +53,10 @@ class MacrocelulaController extends Controller
 
     public function editar($id)
     {
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }
         
         //Pegando informações do usuário que está acessando o sistema 
         $matricula  =  getenv('USERNAME');
@@ -82,6 +90,10 @@ class MacrocelulaController extends Controller
 
     public function incluir(Request $request) 
     {
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }
 
         //Verificando se a solicitação veio da inclusão ou edição
         if($request->input('idmacrocelula') == "") { 
@@ -108,6 +120,10 @@ class MacrocelulaController extends Controller
 
     public function delete($id)
     {
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }
         //Encontrando e deletando Contrato (softDelete)
         $Macrocelula = Macrocelula::find($id);
         $Macrocelula->delete();
