+++ app/Http/Controllers/PrepostoController.php
@@ -16,9 +16,13 @@ class PrepostoController extends Controller
 {
     public function index()
     {
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }
 
         //Pegando informações do usuário que está acessando o sistema
-        $matricula  =   getenv('USERNAME');
+        $matricula = session('matricula');
 
         //Listando as empresas
         $Empresas = Empresa::all();
@@ -43,20 +47,20 @@ class PrepostoController extends Controller
                                  'Empresas'  => $Empresas, 
                                  'Contratos' => $Contratos,
                                  'Prepostos' => $Prepostos,
-        						 'Coordenacoes' => $Coordenacoes,
+                                 'Coordenacoes' => $Coordenacoes,
                                  ]);
 
     }
 
-
-
-
-
     public function editar($id)
     {
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }
         
         //Pegando informações do usuário que está acessando o sistema 
-        $matricula  =  getenv('USERNAME');
+        $matricula = session('matricula');
 
          //Listando as empresas
         $Empresas = Empresa::all();
@@ -83,6 +87,10 @@ class PrepostoController extends Controller
 
     public function incluir(Request $request) 
     {
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }
 
         //Verificando se a solicitação veio da inclusão ou edição
         if($request->input('idpreposto') == "") { 
@@ -110,6 +118,10 @@ class PrepostoController extends Controller
 
     public function delete($id)
     {
+        if (session('matricula') == null) {
+            return view('notificacaoLogin', ['status' => 'expirado']);
+            die();
+        }
         //Encontrando e deletando Contrato (softDelete)
         $Preposto = Preposto::find($id);
         $Preposto->delete();
